requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
langchain>=0.0.267
langchain_community
langchain_openai
